chore: add `pinChangeRequired` field to the user model
This will be used to determine the onboarding path for myCareHub app users.
If a pin change is not required, they will be directed to the home page.
If required, they will go through the whole process of setting up a new
account e.g set security questions, accept terms etc